Move method to shimmed class
